Update to upstream standards
