--- keyboards/keychron/c3_pro/ansi/red/keymaps/keychron/rules.mk
@@ -1,4 +1,0 @@
-VIA_ENABLE = yes
-
-VPATH += keyboards/keychron/common
-SRC += keychron_common.c
