+++ keyboards/keychron/c3_pro/readme.md
@@ -1,20 +1,22 @@
 # Keychron C3 Pro
 
-![Keychron C3 Pro](https://i.imgur.com/79sp3PHh.jpg)
+![Keychron C3 Pro](https://i.imgur.com/b4FTJsf.jpg)
 
 A customizable 87% keyboard.
 
 * Keyboard Maintainer: [Keychron](https://github.com/keychron)
 * Hardware Supported: Keychron C3 Pro
-* Hardware Availability:[Keychron](https://github.com/keychron)
+* Hardware Availability: [Keychron](https://www.keychron.com/)
 
 Make example for this keyboard (after setting up your build environment):
 
     make keychron/c3_pro/ansi/red:default
+    make keychron/c3_pro/ansi/rgb:default 
 
 Flashing example for this keyboard:
 
     make keychron/c3_pro/ansi/red:default:flash
+    make keychron/c3_pro/ansi/rgb:default:flash
 
 ## bootloader
