--- keyboards/keychron/c3_pro_s/ansi/red/red.c
@@ -1,132 +1,0 @@
-/* Copyright 2023 @ Keychron (https://www.keychron.com)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "quantum.h"
-
-#ifdef LED_MATRIX_ENABLE
-// clang-format off
-const snled27351_led PROGMEM g_snled27351_leds[SNLED27351_LED_COUNT] = {
-/* Refer to SNLED27351 manual for these locations
- *   driver
- *   |  LED address
- *   |  | */
-    {0, CB6_CA1},
-    {0, CB6_CA3},
-    {0, CB6_CA4},
-    {0, CB6_CA5},
-    {0, CB6_CA6},
-    {0, CB6_CA7},
-    {0, CB6_CA8},
-    {0, CB6_CA9},
-    {0, CB6_CA10},
-    {0, CB6_CA11},
-    {0, CB6_CA12},
-    {0, CB6_CA13},
-    {0, CB6_CA14},
-    {0, CB6_CA15},
-    {0, CB6_CA16},
-    {0, CB3_CA13},
-
-    {0, CB5_CA1},
-    {0, CB5_CA2},
-    {0, CB5_CA3},
-    {0, CB5_CA4},
-    {0, CB5_CA5},
-    {0, CB5_CA6},
-    {0, CB5_CA7},
-    {0, CB5_CA8},
-    {0, CB5_CA9},
-    {0, CB5_CA10},
-    {0, CB5_CA11},
-    {0, CB5_CA12},
-    {0, CB5_CA13},
-    {0, CB5_CA14},
-    {0, CB5_CA15},
-    {0, CB5_CA16},
-    {0, CB3_CA15},
-
-    {0, CB4_CA1},
-    {0, CB4_CA2},
-    {0, CB4_CA3},
-    {0, CB4_CA4},
-    {0, CB4_CA5},
-    {0, CB4_CA6},
-    {0, CB4_CA7},
-    {0, CB4_CA8},
-    {0, CB4_CA9},
-    {0, CB4_CA10},
-    {0, CB4_CA11},
-    {0, CB4_CA12},
-    {0, CB4_CA13},
-    {0, CB4_CA14},
-    {0, CB4_CA15},
-    {0, CB4_CA16},
-    {0, CB3_CA16},
-
-    {0, CB3_CA1},
-    {0, CB3_CA2},
-    {0, CB3_CA3},
-    {0, CB3_CA4},
-    {0, CB3_CA5},
-    {0, CB3_CA6},
-    {0, CB3_CA7},
-    {0, CB3_CA8},
-    {0, CB3_CA9},
-    {0, CB3_CA10},
-    {0, CB3_CA11},
-    {0, CB3_CA12},
-    {0, CB3_CA14},
-
-    {0, CB2_CA1},
-    {0, CB2_CA3},
-    {0, CB2_CA4},
-    {0, CB2_CA5},
-    {0, CB2_CA6},
-    {0, CB2_CA7},
-    {0, CB2_CA8},
-    {0, CB2_CA9},
-    {0, CB2_CA10},
-    {0, CB2_CA11},
-    {0, CB2_CA12},
-    {0, CB2_CA14},
-    {0, CB2_CA16},
-
-    {0, CB1_CA1},
-    {0, CB1_CA2},
-    {0, CB1_CA3},
-    {0, CB1_CA7},
-    {0, CB1_CA11},
-    {0, CB1_CA12},
-    {0, CB1_CA13},
-    {0, CB1_CA14},
-    {0, CB1_CA15},
-    {0, CB1_CA16},
-    {0, CB2_CA15},
-
-    {0, CB7_CA1},
-    {0, CB7_CA2},
-    {0, CB7_CA3},
-    {0, CB7_CA4},
-    {0, CB7_CA5},
-    {0, CB7_CA6},
-    {0, CB7_CA12},
-    {0, CB7_CA11},
-    {0, CB7_CA10},
-    {0, CB7_CA9},
-    {0, CB7_CA8},
-    {0, CB7_CA7},
-};
-#endif
