--- keyboards/keychron/c3_pro_s/c3_pro.c
@@ -1,126 +1,0 @@
-/* Copyright 2024 @ Keychron (https://www.keychron.com)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "c3_pro.h"
-
-static uint32_t os_switch_timer_buffer   = 0;
-static uint8_t  os_switch_indicate_count = 0;
-
-bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
-    if (!process_record_user(keycode, record)) {
-        return false;
-    }
-    switch (keycode) {
-        case KC_OSSW:
-            if (record->event.pressed) {
-                default_layer_xor(1U << 0);
-                default_layer_xor(1U << 2);
-                eeconfig_update_default_layer(default_layer_state);
-                os_switch_timer_buffer = timer_read32();
-            }
-            return false;
-        default:
-            return true;
-    }
-}
-
-#if defined(RGB_MATRIX_ENABLE) && defined(CAPS_LOCK_LED_INDEX)
-
-bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
-    if (!process_record_user(keycode, record)) { return false; }
-    switch (keycode) {
-#ifdef RGB_MATRIX_ENABLE
-        case QK_RGB_MATRIX_TOGGLE:
-            if (record->event.pressed) {
-                switch (rgb_matrix_get_flags()) {
-                    case LED_FLAG_ALL: {
-                        rgb_matrix_set_flags(LED_FLAG_NONE);
-                        rgb_matrix_set_color_all(0, 0, 0);
-                    } break;
-                    default: {
-                        rgb_matrix_set_flags(LED_FLAG_ALL);
-                    } break;
-                }
-            }
-            if (!rgb_matrix_is_enabled()) {
-                rgb_matrix_set_flags(LED_FLAG_ALL);
-                rgb_matrix_enable();
-            }
-            return false;
-#endif
-    }
-    return true;
-}
-
-bool rgb_matrix_indicators_advanced_kb(uint8_t led_min, uint8_t led_max) {
-    if (!rgb_matrix_indicators_advanced_user(led_min, led_max)) { return false; }
-    // RGB_MATRIX_INDICATOR_SET_COLOR(index, red, green, blue);
-
-    if (host_keyboard_led_state().caps_lock) {
-        RGB_MATRIX_INDICATOR_SET_COLOR(CAPS_LOCK_LED_INDEX, 255, 255, 255);
-    } else {
-        if (!rgb_matrix_get_flags()) {
-            RGB_MATRIX_INDICATOR_SET_COLOR(CAPS_LOCK_LED_INDEX, 0, 0, 0);
-        }
-    }
-    return true;
-}
-
-#endif // RGB_MATRIX_ENABLE && CAPS_LOCK_LED_INDEX
-
-#if defined(LED_MATRIX_ENABLE) && defined(CAPS_LOCK_LED_INDEX)
-
-bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
-    if (!process_record_user(keycode, record)) { return false; }
-    switch (keycode) {
-#ifdef LED_MATRIX_ENABLE
-        case LM_TOGG:
-            if (record->event.pressed) {
-                switch (led_matrix_get_flags()) {
-                    case LED_FLAG_ALL: {
-                        led_matrix_set_flags(LED_FLAG_NONE);
-                        led_matrix_set_value_all(0);
-                    } break;
-                    default: {
-                        led_matrix_set_flags(LED_FLAG_ALL);
-                    } break;
-                }
-            }
-            if (!led_matrix_is_enabled()) {
-                led_matrix_set_flags(LED_FLAG_ALL);
-                led_matrix_enable();
-            }
-            return false;
-#endif
-    }
-    return true;
-}
-
-bool led_matrix_indicators_advanced_kb(uint8_t led_min, uint8_t led_max) {
-    if (!led_matrix_indicators_advanced_user(led_min, led_max)) { return false; }
-
-    if (host_keyboard_led_state().caps_lock) {
-        led_matrix_set_value(CAPS_LOCK_LED_INDEX, 255);
-
-    } else {
-        if (!led_matrix_get_flags()) {
-            led_matrix_set_value(CAPS_LOCK_LED_INDEX, 0);
-        }
-    }
-    return true;
-}
-
-#endif // LED_MATRIX_ENABLE && CAPS_LOCK_LED_INDEX
