--- keyboards/keychron/c3_pro_s/via_json/c3_pro_ansi_red.json
@@ -1,225 +1,0 @@
-{
-  "name": "Keychron C3 Pro ANSI Red",
-  "vendorId": "0x3434",
-  "productId": "0x0430",
-  "keycodes" : ["qmk_lighting"],
-  "customKeycodes": [
-    {"name": "Mission Control", "title": "Mission Control in macOS", "shortName": "MCtrl"},
-    {"name": "Launch Pad", "title": "Launch Pad in macOS", "shortName": "LPad"},
-    {"name": "Left Option", "title": "Left Option in macOS", "shortName": "LOpt"},
-    {"name": "Right Option", "title": "Right Option in macOS", "shortName": "ROpt"},
-    {"name": "Left Cmd", "title": "Left Command in macOS", "shortName": "LCmd"},
-    {"name": "Right Cmd", "title": "Right Command in macOS", "shortName": "RCmd"},
-    {"name": "Siri", "title": "Siri in macOS", "shortName": "Siri"},
-    {"name": "Task View", "title": "Task View in windows", "shortName": "Task"},
-    {"name": "File Explorer", "title": "File Explorer in windows", "shortName": "File"},
-    {"name": "Screen Shot", "title": "Screenshot in macOS", "shortName": "SShot"},
-    {"name": "Cortana", "title": "Cortana in windows", "shortName": "Cortana"},
-    {"name": "OS Switch", "title": "Switch between windows and macOS", "shortName": "OSSW"}
-  ],
-  "matrix": {"rows": 6, "cols": 17},
-  "layouts": {
-    "keymap": [
-      [
-        {
-          "c": "#777777"
-        },
-        "0,0",
-        {
-          "x": 1,
-          "c": "#cccccc"
-        },
-        "0,2",
-        "0,3",
-        "0,4",
-        "0,5",
-        {
-          "x": 0.5,
-          "c": "#aaaaaa"
-        },
-        "0,6",
-        "0,7",
-        "0,8",
-        "0,9",
-        {
-          "x": 0.5,
-          "c": "#cccccc"
-        },
-        "0,10",
-        "0,11",
-        "0,12",
-        "0,13",
-        {
-          "x": 0.25,
-          "c": "#aaaaaa"
-        },
-        "0,14",
-        "0,15",
-        "0,16"
-      ],
-      [
-        {
-          "y": 0.25
-        },
-        "1,0",
-        {
-          "c": "#cccccc"
-        },
-        "1,1",
-        "1,2",
-        "1,3",
-        "1,4",
-        "1,5",
-        "1,6",
-        "1,7",
-        "1,8",
-        "1,9",
-        "1,10",
-        "1,11",
-        "1,12",
-        {
-          "c": "#aaaaaa",
-          "w": 2
-        },
-        "1,13",
-        {
-          "x": 0.25
-        },
-        "1,14",
-        "1,15",
-        "1,16"
-      ],
-      [
-        {
-          "w": 1.5
-        },
-        "2,0",
-        {
-          "c": "#cccccc"
-        },
-        "2,1",
-        "2,2",
-        "2,3",
-        "2,4",
-        "2,5",
-        "2,6",
-        "2,7",
-        "2,8",
-        "2,9",
-        "2,10",
-        "2,11",
-        "2,12",
-        {
-          "c": "#aaaaaa",
-          "w": 1.5
-        },
-        "2,13",
-        {
-          "x": 0.25
-        },
-        "2,14",
-        "2,15",
-        "2,16"
-      ],
-      [
-        {
-          "w": 1.75
-        },
-        "3,0",
-        {
-          "c": "#cccccc"
-        },
-        "3,1",
-        "3,2",
-        "3,3",
-        "3,4",
-        "3,5",
-        "3,6",
-        "3,7",
-        "3,8",
-        "3,9",
-        "3,10",
-        "3,11",
-        {
-          "c": "#777777",
-          "w": 2.25
-        },
-        "3,13"
-      ],
-      [
-        {
-          "c": "#aaaaaa",
-          "w": 2.25
-        },
-        "4,0",
-        {
-          "c": "#cccccc"
-        },
-        "4,2",
-        "4,3",
-        "4,4",
-        "4,5",
-        "4,6",
-        "4,7",
-        "4,8",
-        "4,9",
-        "4,10",
-        "4,11",
-        {
-          "c": "#aaaaaa",
-          "w": 2.75
-        },
-        "4,13",
-        {
-          "x": 1.25,
-          "c": "#777777"
-        },
-        "4,15"
-      ],
-      [
-        {
-          "c": "#aaaaaa",
-          "w": 1.25
-        },
-        "5,0",
-        {
-          "w": 1.25
-        },
-        "5,1",
-        {
-          "w": 1.25
-        },
-        "5,2",
-        {
-          "c": "#cccccc",
-          "w": 6.25
-        },
-        "5,6",
-        {
-          "c": "#aaaaaa",
-          "w": 1.25
-        },
-        "5,10",
-        {
-          "w": 1.25
-        },
-        "5,11",
-        {
-          "w": 1.25
-        },
-        "5,12",
-        {
-          "w": 1.25
-        },
-        "5,13",
-        {
-          "x": 0.25,
-          "c": "#777777"
-        },
-        "5,14",
-        "5,15",
-        "5,16"
-      ]
-    ]
-  }
-}
